+++ src/seminator.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2017, 2019, Fakulta Informatiky Masarykovy univerzity
+// Copyright (C) 2017, 2019, 2020, Fakulta Informatiky Masarykovy univerzity
 //
 // This file is a part of Seminator, a tool for semi-determinization of omega automata.
 //
@@ -21,15 +21,13 @@
 #include <bscc.hpp>
 #include <breakpoint_twa.hpp>
 
-#include <spot/twaalgos/stripacc.hh>
 #include <spot/twaalgos/degen.hh>
 #include <spot/twaalgos/isdet.hh>
 #include <spot/twaalgos/sccinfo.hh>
+#include <spot/twaalgos/minimize.hh>
 #include <spot/misc/optionmap.hh>
 #include <spot/twaalgos/sccfilter.hh>
 #include <spot/twa/bddprint.hh>
-
-static constexpr jobs_type unitjobs[3] = {ViaTGBA, ViaTBA, ViaSBA};
 
 /**
  * Class running possible multiple types of the transformation
@@ -48,7 +46,7 @@ public:
   */
   seminator(spot::twa_graph_ptr input, bool cut_det,
             const spot::option_map* opt = nullptr)
-    : input_(input), opt_(opt), cut_det_(cut_det)
+    : input_(spot::scc_filter(input, true)), opt_(opt), cut_det_(cut_det)
   {
     if (!opt)
       opt_ = new const spot::option_map;
@@ -56,9 +54,12 @@ public:
     preproc_  = opt_->get("preprocess",0);
     postproc_ = opt_->get("postprocess", 1);
 
-    // The deterministic attempt was done in semi_determinize
     if (preproc_)
-      preprocessor_.set_pref(spot::postprocessor::Small);
+      preprocessor_.set_pref(spot::postprocessor::Deterministic);
+    else if (input_->acc().is_all())
+      // automata with "t" acceptance can be determinized and minimized
+      // as a DFA.  (The preprocessor will do that if we use it.)
+      input_ = spot::minimize_monitor(input_);
 
     output_  = static_cast<output_type>(opt_->get("output", TGBA));
 
@@ -81,140 +82,135 @@ public:
   {
     if (jobs == 0)
       jobs = AllJobs;
-    prepare_inputs(jobs);
-    return results_[best_from(jobs)];
-  }
 
-  /**
-  * Run requested jobs
-  *
-  * @param[in] jobs specifies the jobs to run
-  */
-  void run_jobs(jobs_type jobs)
-  {
-    for (auto job : unitjobs)
+    // If the input is already Buchi, there there is no need
+    // for a TGBA job.
+    if (input_->acc().is_buchi() && (jobs & (ViaTBA|ViaSBA)))
       {
-        if (job & jobs)
-          {
-            // Prepare the desired input
-            if (!inputs_[job])
-              prepare_inputs(job);
-            assert(inputs_[job]);
-
-            state_set non_det_states;
-            if (spot::is_deterministic(inputs_[job]) ||
-                is_cut_deterministic(inputs_[job], &non_det_states))
-              results_[job] = inputs_[job];
-            else if (spot::is_semi_deterministic(inputs_[job]))
-              {
-                if (!cut_det_)
-                  results_[job] = inputs_[job];
-                else
-                  results_[job] = determinize_first_component(inputs_[job],
-                                                              &non_det_states);
-              }
-            else
-              {
-                // Run the breakpoint algorithm
-                bp_twa resbp(inputs_[job], cut_det_, opt_);
-                results_[job] = resbp.res_aut();
-                results_[job]->purge_dead_states();
-              }
-
-            // Check the result
-            assert(spot::is_semi_deterministic(results_[job]));
-            assert(!cut_det_ || is_cut_deterministic(results_[job]));
-            assert(results_[job]->acc().is_generalized_buchi());
-
-            if (postproc_)
-              results_[job] = postprocessor_.run(results_[job]);
-
-            switch (output_)
-              {
-              case TBA:
-                results_[job] = spot::degeneralize_tba(results_[job]);
-                if (postproc_)
-                  results_[job] = postprocessor_.run(results_[job]);
-                break;
-              case BA:
-                results_[job] = spot::degeneralize(results_[job]);
-                if (postproc_)
-                  results_[job] = postprocessor_.run(results_[job]);
-                break;
-              default:
-                break;
-              }
-          }
+        jobs &= ~ViaTGBA;
+        // And if the input is already state-based Buchi, there there
+        // is no need for a TBA job.
+        if (input_->prop_state_acc() && (jobs & ViaSBA))
+          jobs &= ~ViaTBA;
       }
-  }
 
-  /**
-  * Choose best from given jobs. Run the algo if not done before.
-  *
-  * @param[in] jobs specifies the jobs to tests
-  */
-    jobs_type best_from(jobs_type jobs)
-    {
-      jobs_type res = 0;
-      for (auto job : unitjobs)
+    spot::twa_graph_ptr best = nullptr;
+    for (auto job : {ViaTGBA, ViaTBA, ViaSBA})
+      if (job & jobs)
         {
-          if (job & jobs)
-            {
-              if (!results_[job])
-                run_jobs(job);
-              if (!res ||
-                  (results_[res]->num_states() > results_[job]->num_states()))
-                res = job;
-            }
+          auto result = postprocess_job(process_job(prepare_input(job)));
+          if (!best || (best->num_states() > result->num_states()))
+            best = result;
         }
-      return res;
-    }
-
+    return best;
+  }
 
 private:
-  /**
-  * Read options from opt into variables
-  */
-  void parse_options();
 
-  /**
-  * Prepare input for requested jobs.
-  *
-  * @param[in] jobs specifies the jobs to run
-  */
-  void prepare_inputs(jobs_type jobs)
+  spot::twa_graph_ptr prepare_input(jobs_type job)
   {
-    // TODO check what makes sense
-    if (jobs & ViaTGBA)
-      inputs_.emplace(ViaTGBA, input_);
-    if (jobs & ViaTBA)
+    switch (job)
       {
-        inputs_.emplace(ViaTBA, spot::degeneralize_tba(input_));
-        if (preproc_)
-          inputs_[ViaTBA] = preprocessor_.run(inputs_[ViaTBA]);
-      }
-    if (jobs & ViaSBA)
-      {
-        inputs_.emplace(ViaSBA, spot::degeneralize(input_));
-        if (preproc_)
+      case ViaTGBA:
+        preprocessor_.set_type(spot::postprocessor::TGBA);
+        return preprocessor_.run(input_);
+      case ViaTBA:
+        {
+          auto out = spot::degeneralize_tba(input_);
+          if (preproc_)
+            {
+              preprocessor_.set_type(spot::postprocessor::TGBA);
+              out = preprocessor_.run(out);
+            }
+          return out;
+        }
+      case ViaSBA:
+        if (!preproc_)
+          {
+            return spot::degeneralize(input_);
+          }
+        else
           {
             preprocessor_.set_type(spot::postprocessor::BA);
-            inputs_[ViaSBA] = preprocessor_.run(inputs_[ViaSBA]);
+            return preprocessor_.run(input_);
           }
+      default:
+        assert(!"should not be reached");
       }
   }
 
+  spot::twa_graph_ptr process_job(spot::twa_graph_ptr input)
+  {
+    spot::twa_graph_ptr result;
+    state_set non_det_states;
+
+    if (spot::is_deterministic(input) ||
+        is_cut_deterministic(input, &non_det_states))
+      {
+        result = input;
+      }
+    else if (spot::is_semi_deterministic(input))
+      {
+        if (!cut_det_)
+          result = input;
+        else
+          result = determinize_first_component(input, &non_det_states);
+      }
+    else
+      {
+        // Run the breakpoint algorithm
+        bp_twa resbp(input, cut_det_, opt_);
+        result = resbp.res_aut();
+        result->purge_dead_states();
+      }
+
+    // Check the result
+    if (!cut_det_)
+      assert(spot::is_semi_deterministic(result));
+    else
+      assert(is_cut_deterministic(result));
+    assert(result->acc().is_generalized_buchi());
+
+    result->prop_semi_deterministic(true);
+    return result;
+  }
+
+  spot::twa_graph_ptr postprocess_job(spot::twa_graph_ptr aut)
+  {
+    if (postproc_)
+      {
+        postprocessor_.set_type(spot::postprocessor::TGBA);
+        aut = postprocessor_.run(aut);
+      }
+
+    switch (output_)
+      {
+      case TBA:
+        aut = spot::degeneralize_tba(aut);
+        if (postproc_)
+          aut = postprocessor_.run(aut);
+        break;
+      case BA:
+        if (postproc_)
+          {
+            postprocessor_.set_type(spot::postprocessor::BA);
+            aut = postprocessor_.run(aut);
+          }
+        else
+          {
+            aut = spot::degeneralize(aut);
+          }
+        break;
+      default:
+        break;
+      }
+    return aut;
+  }
 
   spot::twa_graph_ptr input_;
   const spot::option_map* opt_;
 
-  // Intermediate results
-  typedef std::map<jobs_type, spot::twa_graph_ptr> aut_ptr_dict;
-  aut_ptr_dict inputs_ = aut_ptr_dict();
-  aut_ptr_dict results_ = aut_ptr_dict();
-  spot::twa_graph_ptr best_ = nullptr;
-
-  // Simplifications options (from opt, see parse_options for defaults)
+  // Simplifications options
   bool postproc_;
   bool preproc_;
   bool cut_det_;
@@ -232,44 +228,6 @@ aut_ptr semi_determinize(aut_ptr aut,
                          jobs_type jobs,
                          const_om_ptr opt)
 {
-  if (spot::is_deterministic(aut))
-    return aut;
-
-  bool preproc = false;
-  if (opt)
-    preproc = opt->get("preprocess",0);
-
-  if (preproc)
-  {
-    spot::postprocessor preprocessor;
-    preprocessor.set_pref(spot::postprocessor::Deterministic);
-    preprocessor.run(aut);
-  }
-
-  if (spot::is_semi_deterministic(aut))
-  {
-    if (!cut_det)
-      return aut;
-
-    auto non_det_states = new state_set;
-    aut_ptr result;
-
-    if (is_cut_deterministic(aut, non_det_states))
-      result = aut;
-    else
-      result = determinize_first_component(aut, non_det_states);
-    delete non_det_states;
-    return result;
-  }
-
-  // Safety automata can be determinized using powerset construction
-  if (aut->acc().is_all())
-  {
-    auto result = spot::tba_determinize(aut);
-    result->set_acceptance(0, spot::acc_cond::acc_code::t());
-    return result;
-  }
-
   seminator sem(aut, cut_det, opt);
   return sem.run(jobs);
 }